+++ demo/app/form/datetime-demo.component.ts
@@ -3,7 +3,17 @@ import { Component } from '@angular/core';
 @Component({
   template: `
   <h2>Datetime Component</h2>
+  <p>pop with calendar inside</p>
   <ec-datetime></ec-datetime>
+
+  <h2>Calendar: {{calendar.value}}</h2>
+  <p>controls + month grid inside</p>
+  <ec-calendar #calendar></ec-calendar>
+
+  <h2>Month</h2>
+  <p>Just month grid</p>
+  <ec-month #month></ec-month>
+  {{month.date}}
 `
 })
 export class DatetimeDemoComponent {